Originator/requirements.txt
pycryptodome
requests
Flask==2.1.0
gunicorn==20.1.0